+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/tabs/ManageTabs/header/manage-tabs-header.module.scss
