+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/tabs/ReorderableTabs/reorderable-tabs.module.scss
